Finalize lua aggregation
The script now does aggregation the right way, min aggregation and hour aggregation
are pushed to different queues for processing. Timesamps are correctly rounded

One final flow in the script that doesn't seem to have a solution atm
is that flushing of data to the pollers queues is only forced by the receiving of
new values.

Also metrics now has expire times so we clean up redis if a certain metric is not
updated anymore