improve how AYS repos are loaded and destroy
an repo model is now automaticly created during the loading of the repo
if it doesn't exist yet in db. so we don't force to discover the repo is
we load it manually

don't destroy the full repo bucket in database when we destroy one repo.
Now the repo is properly remove from db and other entries are left
intacts