bug fix concerning the build on 8.2.0
The jsexecutor.json is placed in /optvar/ but this causes a problem because the /optvar is mounted from docker host on to the containers.
This means that all packages outside /opt/jumpscale8/bin such as pip3 packages , apt, or brew. when checked if existing on container will return the value on host instead , causing uninstalled dependencies . To fix this the json file is deleted before each container build to ensure all dependencies are installed.